fix: circular import of request types
